+++ subsys/bluetooth/mesh/prov_device.c
@@ -11,10 +11,6 @@
 #include <sys/util.h>
 #include <sys/byteorder.h>
 
-#include <tinycrypt/constants.h>
-#include <tinycrypt/ecc.h>
-#include <tinycrypt/ecc_dh.h>
-
 #include <net/buf.h>
 #include <bluetooth/bluetooth.h>
 #include <bluetooth/conn.h>
@@ -321,18 +317,14 @@ static void prov_dh_key_gen(void)
 
 	if (IS_ENABLED(CONFIG_BT_MESH_PROV_OOB_PUBLIC_KEY) &&
 	    atomic_test_bit(bt_mesh_prov_link.flags, OOB_PUB_KEY)) {
-		if (uECC_valid_public_key(remote_pk, &curve_secp256r1)) {
-			BT_ERR("Public key is not valid");
-		} else if (uECC_shared_secret(remote_pk, bt_mesh_prov->private_key_be,
-					      bt_mesh_prov_link.dhkey,
-					      &curve_secp256r1) != TC_CRYPTO_SUCCESS) {
-			BT_ERR("DHKey generation failed");
-		} else {
-			dh_key_gen_complete();
+
+		if (bt_mesh_dhkey_gen(remote_pk, bt_mesh_prov->private_key_be,
+				bt_mesh_prov_link.dhkey)) {
+			prov_fail(PROV_ERR_UNEXP_ERR);
 			return;
 		}
 
-		prov_fail(PROV_ERR_UNEXP_ERR);
+		dh_key_gen_complete();
 		return;
 	}
